bigquery/cloud-functions/diagnostics/requirements.txt
google-cloud-storage==1.42.2
google-cloud-bigquery==2.28.1
google-cloud-logging==3.2.1
